Finnished GUI, Started User Manual
User Manual is just a place holder for now. I will popualte it later when I am done. I have gotten the basic functionality of the GUI done. All the important bits are there and now its time to add all the GPIO events and functionality. This if going to be fun to figure out.